refactored Boot code to be able to re-use kamon from tests, updated test suite to run with kamon enabled.
