+++ src/main/scala/ch/chuv/lren/woken/web/BootedCore.scala
@@ -44,6 +44,7 @@ import ch.chuv.lren.woken.dao.{
 import ch.chuv.lren.woken.service._
 import ch.chuv.lren.woken.ssl.WokenSSLConfiguration
 import ch.chuv.lren.woken.backends.woken.WokenClientService
+import ch.chuv.lren.woken.kamon.KamonSupport
 import com.typesafe.scalalogging.LazyLogging
 import kamon.Kamon
 import kamon.prometheus.PrometheusReporter
@@ -72,51 +73,8 @@ trait BootedCore
     with WokenSSLConfiguration
     with LazyLogging {
 
-  override def beforeBoot(): Unit = {
-
-    val kamonConfig = config.getConfig("kamon")
-
-    if (kamonConfig.getBoolean("enabled") || kamonConfig.getBoolean("prometheus.enabled") || kamonConfig
-          .getBoolean("zipkin.enabled")) {
-
-      logger.info("Kamon configuration:")
-      logger.info(config.getConfig("kamon").toString)
-      logger.info(s"Start monitoring...")
-
-      Kamon.reconfigure(config)
-
-      val hostSystemMetrics = kamonConfig.getBoolean("system-metrics.host.enabled")
-      if (hostSystemMetrics) {
-        logger.info(s"Start Sigar metrics...")
-        Try {
-          val sigarLoader = new SigarLoader(classOf[Sigar])
-          sigarLoader.load()
-        }
-
-        Try(
-          SigarProvisioner.provision(
-            new File(System.getProperty("user.home") + File.separator + ".native")
-          )
-        ).recover { case e: Exception => logger.warn("Cannot provision Sigar", e) }
-
-        if (SigarProvisioner.isNativeLoaded)
-          logger.info("Sigar metrics are available")
-        else
-          logger.warn("Sigar metrics are not available")
-      }
-
-      if (hostSystemMetrics || kamonConfig.getBoolean("system-metrics.jvm.enabled")) {
-        logger.info(s"Start collection of system metrics...")
-        SystemMetrics.startCollecting()
-      }
-
-      if (kamonConfig.getBoolean("prometheus.enabled"))
-        Kamon.addReporter(new PrometheusReporter)
-
-      if (kamonConfig.getBoolean("zipkin.enabled"))
-        Kamon.addReporter(new ZipkinReporter)
-    }
-  }
+  override def beforeBoot(): Unit =
+    KamonSupport.startReporters(config)
 
   private lazy val resultsDbConfig = DatabaseConfiguration
     .factory(config)("woken")
