+++ tests/woken-test/src/test/resources/kamon.conf
@@ -3,7 +3,8 @@ kamon {
   enabled = ${?KAMON_ENABLED}
 
   environment {
-    service = "woken"
+    service = "woken-test"
+    service = ${?KAMON_SERVICE_NAME}
   }
 
   zipkin = {
